[add] add arguments TOKEN and MAIL for coverity automated use. add coverity tool scripted retrieval. add coverity automated upload

+++ nodegl-sh-ci.sh
@@ -1,5 +1,35 @@
 #!/bin/bash
 
+set -e
+
+printHelp() {
+  echo "Usage: nodegl-sh-ci.sh <coverity token> <email>"
+  echo ""
+}
+
+if [ $# -eq 0 ]; then
+  printHelp
+  exit 0
+fi
+if [ $# -ne 2 ]; then
+  echo "ERROR: wrong number of arguments"
+  printHelp
+  exit 0
+fi
+
+TOKEN=$1
+MAIL=$2
+
+# Verify parameters
+if [ -z "${MAIL}" ] ; then
+  echo "MAIL is not specified"
+  exit 1;
+fi 
+if [ -z "${TOKEN}" ] ; then
+  echo "TOKEN is not specified"
+  exit 1;
+fi 
+
 # clean previous results
 rm -rf ./cov-int.tgz ./test-results
 
@@ -18,11 +48,25 @@ if [[ -z "$COVERITY_PATH" ]]; then
 	echo "Docker volume 'coverity' not found. It will de be created"
 	docker volume create coverity
 	COVERITY_PATH=$(docker volume inspect --format '{{ .Mountpoint }}' coverity)
-	echo "untar coverity into volume 'coverity' "
-	# untar coverity into coverity volume
-	tar -xvzf cov.tgz -C $COVERITY_PATH
+	# get coverity tarball
+	wget https://scan.coverity.com/download/linux64 --post-data "token=$TOKEN&project=node.gl" -O coverity_tool.tgz
+	# get tarbell hash
+	wget https://scan.coverity.com/download/linux64 --post-data "token=$TOKEN&project=node.gl&md5=1" -O coverity_tool.md5
+	# verify coverity download
+	if[ "$(md5sum /home/path/file1.txt |awk '{print $1)')" == "$(md5sum /home/path/file2.txt |awk '{print $1}')" ]; then
+		echo "untar coverity into volume 'coverity' "
+		# untar coverity into coverity volume
+		tar -xvzf coverity_tool.tgz -C $COVERITY_PATH
+		echo "Docker volume 'coverity' created at $COVERITY_PATH."
+	else
+		docker volume rm coverity
+		echo "coverity download failed"
+		exit 1
+	fi
+else
+	echo "Docker volume 'coverity' found at $COVERITY_PATH."
 fi
-echo "Docker volume 'coverity' found at $COVERITY_PATH."
+
 
 
 # get "build-script" docker volume path; create it if it does not exists yet
@@ -40,7 +84,7 @@ echo "Docker volume 'build-script' found at $BUILD_SCRIPT_PATH."
 
 # start docker container used for build and running tests
 echo "start container"
-docker run -d --name nodegl-build -t --mount source=coverity,destination=/home/root/coverity --mount source=build-script,destination=/home/root/build-script nodegl:CI
+docker run -d --name nodegl-build -t --mount source=coverity,destination=/home/root/coverity --mount source=build-script,destination=/home/root/build-script nodegl:CI -e TOKEN=$TOKEN -e MAIL=$MAIL
 
 # launch build and tests
 echo "launch build and test on container"
@@ -51,6 +95,15 @@ echo "recover test results artifacts"
 cp -R $COVERITY_PATH/cov-int .
 cp -R $BUILD_SCRIPT_PATH/tests-results .
 tar cvzf cov-int.tgz cov-int
+
+# upload results
+echo "upload result to coverity scan
+curl --form token=$TOKEN \
+  --form email=$MAIL \
+  --form file=@cov-int.tgz \
+  --form version="Version" \
+  --form description="Description" \
+  https://scan.coverity.com/builds?project=node.gl
 
 # clean for next run
 echo "clean build artifacts"
